% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confidence.R
\name{confidence}
\alias{confidence}
\title{confidence}
\usage{
confidence(s, type)
}
\arguments{
\item{s}{the score}

\item{type}{coding or noncoding}
}
\value{
returns a confidence level or remark/error message
}
\description{
This function annotated a confidence level to the score
}
\examples{

remark <- confidence(s, type)
}
