% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rdata.R
\docType{data}
\name{cscape_somatic_output}
\alias{cscape_somatic_output}
\title{Cscape-somatic annotations of TCGA-LUAD}
\format{
A 645x7 matrix.
}
\usage{
data(cscape_somatic_output)
}
\value{
A 645x7 matrix.
}
\description{
Output from DMA. 
This contains the cscape-somatic annotations for all differentially expressed genes
}
\keyword{datasets}
