% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rdata.R
\docType{data}
\name{DAPsmatrix}
\alias{DAPsmatrix}
\title{Differentially abundant proteins}
\format{
A 4873x4 matrix
}
\usage{
data(DAPsmatrix)
}
\value{
A 4873x4 matrix
}
\description{
A matrix containing differentially abundant proteins between colon adenocarinoma cancer and normal
samples found using CPTAC data portal.
}
\details{
The matrix contains the differentially abundant protein with gene names in rows and log2 fold change
PValues and  Direction of abundance change values in columns.
}
\keyword{datasets}
