% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rdata.R
\docType{data}
\name{dataDMA}
\alias{dataDMA}
\title{Output example from the function Driver Mutation Analysis}
\format{
A list of two.
}
\usage{
data(dataDMA)
}
\value{
A list of two, containing 0 tumor-suppressor and 1 oncogene.
}
\description{
The predicted driver genes, which have at least one driver mutation.
}
\keyword{datasets}
