% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rdata.R
\docType{data}
\name{dataEDA}
\alias{dataEDA}
\title{Output example from EDA function}
\format{
A list of length two
}
\usage{
data(dataEDA)
}
\value{
A list of length two
}
\description{
The predicted driver genes based on methylation
evidence
}
\details{
The data is a list of two elements where each 
element represents predicted oncogenes and tumor suppressors
}
\keyword{datasets}
