% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rdata.R
\docType{data}
\name{dataFEA}
\alias{dataFEA}
\title{Functional enrichment analysis with ORA method}
\format{
A dataframe of dimension 101x7
}
\usage{
data(dataFEA)
}
\value{
A dataframe of dimension 101x7
}
\description{
The output of the FEA function which does enrichment analysis using ORA-based method
}
\details{
The input to the FEA is the differentially expressed genes.
}
\keyword{datasets}
