% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rdata.R
\docType{data}
\name{dataFilt}
\alias{dataFilt}
\title{Gene expression data from TCGA-LUAD}
\format{
A 3000x20 matrix
}
\usage{
data(dataFilt)
}
\value{
A 3000x20 matrix
}
\description{
A matrix that provides processed gene expression 
data (obtained from RNA seq) from the TCGA-LUAD project
}
\details{
The matrix contains the genes in rows and samples in 
columns. The data has been downloaded and processed using 
TCGAbiolinks.
}
\keyword{datasets}
