% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rdata.R
\docType{data}
\name{dataGLS}
\alias{dataGLS}
\title{Literature search of driver genes}
\format{
A 13x8 tibble.
}
\usage{
data(dataGLS)
}
\value{
A 13x8 tibble.
}
\description{
A tibble containing results of literature search where predicted
driver genes stored in dataDMA were queried for their role as drivers in
PubMed
}
\details{
The tibble contains PubMed IDs, doi, title, abstract, year of publication,
keywords, and total number of publications for the genes.
}
\keyword{datasets}
