% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rdata.R
\docType{data}
\name{dataGRN_no_noise}
\alias{dataGRN_no_noise}
\title{Gene regulatory network}
\format{
A list of 2 elements where the first element is a 23x613 matrix 
and the second element is a vector of length 23
}
\usage{
data(dataGRN_no_noise)
}
\value{
A list of 2 elements where the first element is a 23x613 matrix and the second element is a vector of length 23
}
\description{
The output of the GRN function which finds connections 
between genes where the noise is set to 0 for testing reproducibility purposes.
}
\details{
The input to the GRN is the differentially expressed genes and the gene expression data.
}
\keyword{datasets}
