% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rdata.R
\docType{data}
\name{dataGTM}
\alias{dataGTM}
\title{Text mining of candidate driver genes}
\format{
A 14x4 tibble.
}
\usage{
data(dataGTM)
}
\value{
A 14x4 tibble.
}
\description{
A tibble containing results of text mining where predicted
driver genes stored in dataDMA were queried for their role as drivers in
PubMed
}
\details{
The tibble contains the PubMed IDs for a given gene in rows and
the columns contain the given gene, the query, total number of publications,
and PubMed ID.
}
\keyword{datasets}
