% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rdata.R
\docType{data}
\name{dataMAF}
\alias{dataMAF}
\title{Mutation data from TCGA LUAD}
\format{
A 3000x141 matrix.
}
\usage{
data(dataMAF)
}
\value{
A 3000x141 matrix.
}
\description{
An examplary MAF file from TCGA on lung cancer LUAD. It contains 3000 randomly selected mutations.
}
\keyword{datasets}
