% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rdata.R
\docType{data}
\name{dataMAVISp}
\alias{dataMAVISp}
\title{MAVISp dataset example}
\format{
A list of data frames (183 x 43 and 1180 x 46).
}
\usage{
data(dataMAVISp)
}
\value{
A list of tibbles (183 x 43 and 1180 x 46).
}
\description{
MAVISp results for proteins TP53 and RUNX1, simple mode. They are used
for testing and vignettes. They are stored as a list of tibbles.
}
\details{
The dataframes contain the full available MAVISp data for simple mode at
present time.
}
\keyword{datasets}
