% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rdata.R
\docType{data}
\name{dataMethyl}
\alias{dataMethyl}
\title{Methylation data matrix from TCGA-LUAD project}
\format{
A 73x27 matrix.
}
\usage{
data(dataMethyl)
}
\value{
A 73x27 matrix.
}
\description{
A data matrix containing methylation data from TCGA-LUAD
where CpG probes are in rows and samples are in columns.
}
\details{
The CpG probes are in rows and samples are in columns.
}
\keyword{datasets}
