% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rdata.R
\docType{data}
\name{dataPRA}
\alias{dataPRA}
\title{Output example from function Pattern Recognition Analysis}
\format{
A list of two.
}
\usage{
data(dataPRA)
}
\value{
A list of two.
}
\description{
The predicted TSGs and OCGs and their moonlight gene z-score.
The PRA() were run with expert-based approach with apoptosis and proliferation of cells.
}
\keyword{datasets}
