% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rdata.R
\docType{data}
\name{dataURA}
\alias{dataURA}
\title{Upstream regulator analysis}
\format{
A 100x2 matrix
}
\usage{
data(dataURA)
}
\value{
A 100x2 matrix
}
\description{
The output of the URA function which carries out the 
upstream regulator analysis
}
\details{
The input to URA is the output of GRN and a list of biological 
processes and the differentially expressed genes
}
\keyword{datasets}
