% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rdata.R
\docType{data}
\name{dataURA_plot}
\alias{dataURA_plot}
\title{Upstream regulator analysis}
\format{
A 12x2 matrix
}
\usage{
data(dataURA_plot)
}
\value{
A 12x2 matrix
}
\description{
The output of the URA function which carries out the 
upstream regulator analysis
}
\details{
This URA data is used to showcase some of the visualization
functions
}
\keyword{datasets}
