% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rdata.R
\docType{data}
\name{DEG_Methylation_Annotations}
\alias{DEG_Methylation_Annotations}
\title{Output example from GMA function}
\format{
A 3435x35 tibble
}
\usage{
data(DEG_Methylation_Annotations.csv)
}
\value{
A 3435x35 tibble
}
\description{
Output file from running GMA function which
is a summary of DEGs and associated CpG probes
}
\details{
The data is a table where each row is a CpG probe
in a DEG. Various annotations such as start/end site of
CpG probe, promoter/enhancer annotations, NCG annotations
are included in the table.
}
\keyword{datasets}
