% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rdata.R
\docType{data}
\name{DEG_Mutations_Annotations}
\alias{DEG_Mutations_Annotations}
\title{Differentially expressed genes's Mutations}
\format{
A 3561x173 matrix.
}
\usage{
data(DEG_Mutations_Annotations)
}
\value{
A 3561x173 matrix.
}
\description{
Output from DMA. 
This contains the differentially expressed genes's mutations and all annotations
generated in DMA() on the TCGA-LUAD project.
}
\keyword{datasets}
