% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rdata.R
\docType{data}
\name{DEGsmatrix}
\alias{DEGsmatrix}
\title{Differentially expressed genes}
\format{
A 3390x5 matrix
}
\usage{
data(DEGsmatrix)
}
\value{
A 3390x5 matrix
}
\description{
A matrix containing differentially expressed genes between lung cancer and normal
samples found using TCGA-LUAD data and TCGAbiolinks.
}
\details{
The matrix contains the differentially expressed genes in rows and log2 fold change
and FDR values in columns.
}
\keyword{datasets}
