% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rdata.R
\docType{data}
\name{DiseaseList}
\alias{DiseaseList}
\title{Cancer-related biological processes}
\format{
A list of 101 elements
}
\usage{
data(DiseaseList)
}
\value{
A list of 101 elements
}
\description{
A dataset containing information about 101 cancer-related biological processes.
}
\details{
The dataset contains a list of the 101 biological processes which includes genes
playing a role in each biological processes including literature findings of the genes' function
in the biological processes.
}
\keyword{datasets}
