% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rdata.R
\docType{data}
\name{EAGenes}
\alias{EAGenes}
\title{Information about genes}
\format{
A 20038x5 matrix
}
\usage{
data(EAGenes)
}
\value{
A 20038x5 matrix
}
\description{
A matrix containing information about 20038 genes 
including their gene description, location and family
}
\details{
The matrix contains the genes in rows and description, 
location and family in columns.
}
\keyword{datasets}
