% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rdata.R
\docType{data}
\name{EncodePromoters}
\alias{EncodePromoters}
\title{Promoters}
\format{
A tibble with no columnnames or rownames.
\enumerate{
\item The first column is chromosome eg. chr1
\item The second column is start position eg. 10451
\item The third column is end position eg. 10563
}
}
\source{
\url{https://www.encodeproject.org/}
}
\usage{
data(EncodePromoters)
}
\value{
A 84738x6 table
}
\description{
Experimentially verified promoter sites by J. Michael Cherry, Stanford.
Downloaded from the ENCODE identifier ENCSR294YNI.
It contains chromosome, start and end sites of promoters.
}
\references{
ENCODE identifier: ENCSR294YNI

Luo Y, Hitz BC, Gabdank I, Hilton JA, Kagda MS, Lam B, Myers Z, Sud P, Jou J,
Lin K, Baymuradov UK, Graham K, Litton C, Miyasato SR, Strattan JS, Jolanki O,
Lee JW, Tanaka FY, Adenekan P, O'Neill E, Cherry JM.
New developments on the Encyclopedia of DNA Elements (ENCODE) data portal.
Nucleic Acids Res. 2020 Jan 8;48(D1):D882-D889. doi: 10.1093/nar/gkz1062.
PMID: 31713622; PMCID: PMC7061942.
}
\keyword{datasets}
