% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EpiMix_getInfiniumAnnotation.R
\name{EpiMix_getInfiniumAnnotation}
\alias{EpiMix_getInfiniumAnnotation}
\title{EpiMix_getInfiniumAnnotation
This function gets Infinium probe annotation from the
sesameData library. This function is from the EpiMix package
https://bioconductor.org/packages/release/bioc/html/EpiMix.html.
Zheng Y, Jun J, Gevaert O (2023). EpiMix: 
EpiMix: an integrative tool for the population-level analysis of 
DNA methylation. R package version 1.1.2.}
\usage{
EpiMix_getInfiniumAnnotation(plat = "EPIC", genome = "hg38")
}
\arguments{
\item{plat}{A character string representing the methylation platform 
which can either be HM27, HM450 or EPIC}

\item{genome}{A character string representing the genome build version
which can either be hg19 or hg38}
}
\value{
Probe annotations
}
\description{
EpiMix_getInfiniumAnnotation
This function gets Infinium probe annotation from the
sesameData library. This function is from the EpiMix package
https://bioconductor.org/packages/release/bioc/html/EpiMix.html.
Zheng Y, Jun J, Gevaert O (2023). EpiMix: 
EpiMix: an integrative tool for the population-level analysis of 
DNA methylation. R package version 1.1.2.
}
\keyword{internal}
