% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rdata.R
\docType{data}
\name{EpiMix_Results_Regular}
\alias{EpiMix_Results_Regular}
\title{Output example from GMA function}
\format{
A list of length nine
}
\usage{
data(EpiMix_Results_Regular)
}
\value{
A list of length nine
}
\description{
The object, a list, that was returned from
running the EpiMix function and is one of the outputs from the 
GMA function.
}
\details{
The data is a list of nine elements which is outputted
from the EpiMix function
}
\keyword{datasets}
