% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rdata.R
\docType{data}
\name{GEO_TCGAtab}
\alias{GEO_TCGAtab}
\title{Information on GEO and TCGA data}
\format{
A 18x12 matrix
}
\usage{
data(GEO_TCGAtab)
}
\value{
A 18x12 matrix
}
\description{
A matrix that provides the GEO dataset
  matched to one of 18 TCGA cancer types
}
\details{
The matrix contains the cancer types in rows and 
information about sample type from both TCGA and GEO in columns.
}
\keyword{datasets}
