% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GTM.R
\name{GTM}
\alias{GTM}
\title{GTM
This function carries out gene text mining.}
\usage{
GTM(genes, query_string = "AND cancer AND driver", max_records = 20)
}
\arguments{
\item{genes}{A character string containing the genes
to perform text mining using PubMed database}

\item{query_string}{A character string containing words 
in query to follow the gene of interest. Default is 
"AND cancer AND driver" resulting in a final query of
"Gene AND cancer AND driver". Standard PubMed syntax
can be used in the query. For example Boolean operators
AND, OR, NOT can be applied and tags such as [AU], 
[TITLE/ABSTRACT], [Affiliation] can be used.}

\item{max_records}{An integer containing the maximum
number of records to be fetched from PubMed.}
}
\value{
A tibble containing results of text mining where
PubMed was queried for information of input genes. Each
row in the tibble contains a sentence from an abstract
that contains either of the words supplied in the query.
Each row also contains corresponding PubMed ID, gene, 
total number of PubMed publications, doi, title, full
abstract and year of publications, resulting in a total
of eight columns.
}
\description{
GTM
This function carries out gene text mining.
}
\examples{
data("dataDMA") 
genes_query <- Reduce(c, dataDMA)
dataGTM <- GTM(genes = genes_query) 
}
