% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rdata.R
\docType{data}
\name{knownDriverGenes}
\alias{knownDriverGenes}
\title{Information of known cancer driver genes from COSMIC}
\format{
A list containing two elements where the first element is a 
character vector of 55 and the second element is a character vector of #' 84
}
\usage{
data(knownDriverGenes)
}
\value{
A list containing two elements where the first element is a 
character vector of 55 and the second element is a character vector of #' 84
}
\description{
A list of known cancer driver genes from COSMIC
}
\details{
The list contains two elements: a vector of known tumor #' suppressors and a vector of known oncogenes
}
\keyword{datasets}
