% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LiftMAF.R
\name{LiftMAF}
\alias{LiftMAF}
\title{LiftMAF}
\usage{
LiftMAF(Infile, Current_Build)
}
\arguments{
\item{Infile}{A tibble of MAF.}

\item{Current_Build}{A charcter string, either \code{GRCh38} or \code{GRCh37}}
}
\value{
MAF tibble with positions lifted to another build
}
\description{
This function lifts a MAF file to a different genomic build.
}
\examples{
data(dataMAF)
dataMAF_example <- dataMAF[1,]
LiftMAF(dataMAF_example, Current_Build = 'GRCh38')
}
