% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rdata.R
\docType{data}
\name{listMoonlight}
\alias{listMoonlight}
\title{List of oncogenic mediators of 5 TCGA cancer types}
\format{
A list containing 5 elements where each element contains 
differentially expressed genes and output from the URA and PRA 
functions of 5 TCGA cancer types
}
\usage{
data(listMoonlight)
}
\value{
A list containing 5 elements where each element contains 
differentially expressed genes and output from the URA and PRA 
functions of 5 TCGA cancer types
}
\description{
A list of oncogenic mediators of 5 TCGA cancer types: 
BLCA, BRCA, LUAD, READ and STAD
}
\details{
Each element in the list contains differentially expressed 
genes and output from the URA and PRA functions
}
\keyword{datasets}
