% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rdata.R
\docType{data}
\name{LOC_translation}
\alias{LOC_translation}
\title{Level of Consequence: Translation}
\format{
A 18x7 table
}
\usage{
data(LOC_translation)
}
\value{
A 18x7 table
}
\description{
A dataset describing if a mutation of a certain class and
type possibly have an effect on peptide level.
}
\details{
The values are binary: 0 no effect is possible, 1 an effect is possible.

See supplementary material for details.
}
\references{
paper
}
\keyword{datasets}
