% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rdata.R
\docType{data}
\name{LUAD_sample_anno}
\alias{LUAD_sample_anno}
\title{Sample annotations of TCGA-LUAD project}
\format{
A 23x2 matrix
}
\usage{
data(LUAD_sample_anno)
}
\value{
A 23x2 matrix
}
\description{
A matrix that annotates LUAD samples as either
cancer or normal
}
\details{
The matrix contains two columns: "primary" which
contains patient barcodes of TCGA-LUAD and "sample.type"
which denotes if the sample is either a "Cancer" or 
"Normal" sample
}
\keyword{datasets}
