% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rdata.R
\docType{data}
\name{MetEvidenceDriver}
\alias{MetEvidenceDriver}
\title{Methylation evidence table to define driver genes}
\format{
A 30x6 tibble.
}
\usage{
data(MetEvidenceDriver)
}
\value{
A 30x6 tibble.
}
\description{
A tibble containing combinations of methylation states
of probes used to define driver genes
}
\details{
The tibble contains a value of 1 if a probe is found that
is either hypo-, hyper-, dualmethylated or not methylated. This is
compared with Moonlight's predictions of role of oncogenic mediators
to define driver genes based on methylation evidence.
}
\keyword{datasets}
