% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Moonlight2R.R
\docType{package}
\name{Moonlight2R}
\alias{Moonlight2R}
\title{Moonlight2R}
\description{
Moonlight2R is a package designed for the identification of cancer driver genes. Please see the
documentation on our Bioconductor page for more details: https://www.bioconductor.org/packages/release/bioc/html/MoonlightR.html

If you experience issues with the package, please open an Issue on our GitHub repository: https://github.com/ELELAB/Moonlight2R

If you use this package in your research, please cite this paper: https://doi.org/10.1038/s41467-019-13803-0
}
