% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rdata.R
\docType{data}
\name{NCG}
\alias{NCG}
\title{Network of Cancer Genes 7.0}
\format{
The format have been rearranged from the original.
<symbold>|<NCG_driver>|<NCG_cgc_annotation>|<NCG_vogelstein_annotation>|
<NCG_saito_annotation>|<NCG_pubmed_id>
}
\source{
\url{http://ncg.kcl.ac.uk/}
}
\usage{
data(NCG)
}
\value{
A 3347x7 table
}
\description{
A dataset retrived from Network of Cancer Genes 7.0
}
\details{
The NCG_driver is reported as a OCG or TSG when at least one of three
three databases have documented it. These are cosmic gene census (cgc),
vogelstein et al. 2013 or saito et al. 2020. The NCG_driver is reported as a
candidate, when literature support the gene as a cancer driver.
}
\references{
Comparative assessment of genes driving cancer and somatic
evolution in non-cancer tissues: an update of the Network of Cancer Genes (NCG)
resource.
Dressler L., Bortolomeazzi M., Keddar M.R., Misetic H., Sartini G.,
Acha-Sagredo A., Montorsi L., Wijewardhane N., Repana D., Nulsen J.,
Goldman J., Pollit M., Davis P., Strange A., Ambrose K. and Ciccarelli F.D.
}
\keyword{datasets}
