% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rdata.R
\docType{data}
\name{Oncogenic_mediators_methylation_summary}
\alias{Oncogenic_mediators_methylation_summary}
\title{Output example from EDA function}
\format{
A 25x7 tibble
}
\usage{
data(Oncogenic_mediators_methylation_summary)
}
\value{
A 25x7 tibble
}
\description{
Output file from running the EDA function which
is a summary of the oncogenic mediators and their sum of
methylated CpG probes together with the evidence level of
their role as driver gene.
}
\details{
The data is a table where each row is an oncogenic
mediator and the columns represent the predicted driver role
and the sum of hypo-, hyper-, and dualmethylated CpG sites.
}
\keyword{datasets}
