% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rdata.R
\docType{data}
\name{Oncogenic_mediators_mutation_summary}
\alias{Oncogenic_mediators_mutation_summary}
\title{Oncogenic Mediators Mutation Summary}
\format{
A 1105x15 matrix.
}
\usage{
data(Oncogenic_mediators_mutation_summary)
}
\value{
A 1105x15 matrix.
}
\description{
Output from DMA. 
This contains the oncogenic mediator from the TCGA-LUAD project, and their mutation
assesments summarised based on CSCape-somatic and Level of Consequence.
}
\keyword{datasets}
