% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotHeatmap.R
\name{plotHeatmap}
\alias{plotHeatmap}
\title{plotHeatmap}
\usage{
plotHeatmap(df)
}
\arguments{
\item{df}{a tibble}
}
\value{
The name of the alphabeatically first gene in the tibble
}
\description{
This function creates a unclustered heatmap from the inputted data tibble
and saves it
}
