% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PRAtoTibble.R
\name{PRAtoTibble}
\alias{PRAtoTibble}
\title{PRAtoTibble}
\usage{
PRAtoTibble(pra_file)
}
\arguments{
\item{pra_file}{RDS object (list of two) from PRA}
}
\value{
tibble with drivers
}
\description{
This function changes the PRA output to tibble format
}
\examples{

PRAtoTibble(pra_file)
}
