% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunCscape_somatic.R
\name{RunCscape_somatic}
\alias{RunCscape_somatic}
\title{RunCscape_somatic}
\usage{
RunCscape_somatic(input, coding_file, noncoding_file)
}
\arguments{
\item{input}{Input matching cscape input}

\item{coding_file}{cscape_table with coding scores}

\item{noncoding_file}{cscape_table with noncoding scores}
}
\value{
returns a tibble with a score and remark for each SNP
}
\description{
This function retrive cscape-scores to SNPs
}
\examples{
\donttest{
cscape_out <- RunCscape_somatic(input, coding_file, noncoding_file)
}
}
