% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rdata.R
\docType{data}
\name{tabGrowBlock}
\alias{tabGrowBlock}
\title{Information of growing/blocking characteristics of 101 biological
processes}
\format{
A 101x3 matrix
}
\usage{
data(tabGrowBlock)
}
\value{
A 101x3 matrix
}
\description{
A matrix with biological processes in rows and the cancer #' growing or blocking effect of the process in columns
}
\details{
For each biological processes the cancer growing/blocking 
effect is indicated
}
\keyword{datasets}
