% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabix_func.R
\name{tabix_func}
\alias{tabix_func}
\title{tabix_func}
\usage{
tabix_func(Ranges, Reference_Allele, Mutant, file_coding, file_noncoding)
}
\arguments{
\item{Ranges}{The position}

\item{Reference_Allele}{The reference nucleotide}

\item{Mutant}{The mutant nucleotide}

\item{file_coding}{cscape_table with coding scores}

\item{file_noncoding}{cscape_table with noncoding scores}
}
\value{
returns the score
}
\description{
This function retrives the individial score for a SNP
}
\examples{
\dontrun{ 
data <- tabix_func(Ranges, Reference_Allele, Mutant, file_coding, file_noncoding)
}
}
