% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rdata.R
\docType{data}
\name{dataFEA_fgsea}
\alias{dataFEA_fgsea}
\title{Functional enrichment analysis with fgsea method}
\format{
A dataframe of dimension 101x8
}
\usage{
data(dataFEA_fgsea)
}
\value{
A dataframe of dimension 101x8
}
\description{
The output of the FEA function which does enrichment analysis using fgsea method
}
\details{
The input to the FEA is the differentially expressed genes.
}
\keyword{datasets}
