% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rdata.R
\docType{data}
\name{dataGTM}
\alias{dataGTM}
\title{Text mining of candidate driver genes}
\format{
A 57x8 tibble.
}
\usage{
data(dataGTM)
}
\value{
A 57x8 tibble.
}
\description{
A tibble containing results of text mining where predicted
driver genes stored in dataDMA were queried for their role as drivers in
PubMed
}
\details{
The tibble contains sentences from abstracts from PubMed IDs 
that contain either of the words supplied in the query. The tibble also
contains PubMed IDs, total number of publications, doi, title, full
abstract and year of publication.
}
\keyword{datasets}
