% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DMA.R
\name{DMA}
\alias{DMA}
\title{DMA}
\usage{
DMA(
  dataMAF,
  dataDEGs,
  dataPRA,
  runCscape = TRUE,
  coding_file,
  noncoding_file,
  results_folder = "./DMAresults"
)
}
\arguments{
\item{dataDEGs}{Output DEA function.}

\item{dataPRA}{Output PRA function.}

\item{runCscape}{Bolean. If \code{FALSE} will load CScape output file from results-folder
Default = \code{TRUE}.}

\item{coding_file}{A character string. Path to and name of CScape-somatic coding file.
Can be downloaded at \url{http://cscape-somatic.biocompute.org.uk/#download}. The .tbi file must be placed in the same folder.}

\item{noncoding_file}{A charcter string. Path to and name of CScape-somatic noncoding file.
Can be downloaded at \url{http://cscape-somatic.biocompute.org.uk/#download}. The .tbi file must be placed in the same folder.}

\item{results_folder}{A character string. Path to the results generated by this function.}

\item{dataMaf}{A MAF file rda object.
The MAF file must at least contain the following columns
\itemize{
\item Hugo_Symbol eg. BRCA1
\item Chromosome eg. chr1
\item Start_Position eg. 54402
\item End_Position e.g. 54443
\item Strand eg. +
\item Variant_Classification
\item Variant_Type
\item Reference_Allele
\item Tumor_Seq_Allele1
\item Tumor_Seq_Allele2
}}
}
\value{
List of two, containing TSGs and OCGs with at least one driver mutation. Additionally files are saved in \code{results_folder}.
}
\description{
This function carries out the driver mutation analysis
}
\examples{

DMA(dataMAF = dataMAF,
    dataDEGs = DEGmatrix,
    dataPRA = dataPRA,
    coding_file = "path/css_coding.vcf.gz",
    noncoding_file = "path/css_noncoding.vcf.gz",
    results_folder = "path/results")

#If the cscape-somatic file have already been created
cscape_somatic_output <- read.csv("./results/Cscape_somatic_output.csv")
save(cscape_somatic_output, file = "./results/Cscape_somatic_output.rda")
    
DMA(dataMAF = dataMAF,
    dataDEGs = DEGmatrix,
    dataPRA = dataPRA,
    runCscape = FALSE,
    results_folder = "./results")     
}
