% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FEA.R
\name{FEA}
\alias{FEA}
\title{FEA}
\usage{
FEA(BPname = NULL, DiffMatrix, method = "ora", seed = 123)
}
\arguments{
\item{BPname}{BPname biological process such as "proliferation of cells", "ALL" (default) if FEA should be carried out for all 101 biological processes}

\item{DiffMatrix}{output from DEA such as dataDEGs or from DAA such as differential abundace matrix}

\item{method}{method to perform Functional Enirchment analysis. Must be either "ora" (Over-Representation Analysis) or "fgsea" (Fast Gene Set Enrichment Analysis)}

\item{seed}{internal seed before the 'fgsea' function is called to ensure reproducibility of the permutation-based p-values. Default is 123}
}
\value{
matrix from FEA
}
\description{
This function carries out the functional enrichment analysis (FEA).
The analysis can be performed using as input a data matrix from Differential Expression Analysis of genes or Differential Abundance Analysis of proteins.
The analysis can be performed using the ORA (Over Representation Analysis)-based method which implements Fisher's Exact Test or the pre-ranked FCS(Functional Class Scoring)-based method
which implements fgsea (Fast Gene Set Enrichment Analysis).
}
\examples{
data(DEGsmatrix)
data(DiseaseList)
data(EAGenes)
DEGsmatrix <- DEGsmatrix[seq.int(2), ]
dataFEA <- FEA(DiffMatrix = DEGsmatrix, BPname = "apoptosis")
}
