% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GTM.R
\name{GTM}
\alias{GTM}
\title{GTM
This function carries out gene text mining.}
\usage{
GTM(genes, query_string = "AND cancer AND driver")
}
\arguments{
\item{genes}{A character string containing the genes
to perform text mining using PubMed database}

\item{query_string}{A character string containing words 
in query to follow the gene of interest. Default is 
"AND cancer AND driver" resulting in a final query of
"Gene AND cancer AND driver". Standard PubMed syntax
can be used in the query. For example Boolean operators
AND, OR, NOT can be applied and tags such as [AU], 
[TITLE/ABSTRACT], [Affiliation] can be used.}
}
\value{
A tibble containing results of text mining where
PubMed was queried for information of input genes. Each
row in the tibble contains a PubMed ID for a given gene.
If more than 20 results from the specified query is obtained, 
only the first 20 are included. The tibble contains four
columns: Hugo_Symbol which contains the given gene, 
pubmed_query which contains the query that was searched,
pubmed_count which is the number of total publications 
that the query resulted in and pubmed_ids which contains the
PubMed IDs.
}
\description{
GTM
This function carries out gene text mining.
}
\examples{
data("dataDMA") 
genes_query <- Reduce(c, dataDMA)
dataGTM <- GTM(genes = genes_query) 
}
