% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadMAVISp.R
\name{loadMAVISp}
\alias{loadMAVISp}
\title{loadMAVISp}
\usage{
loadMAVISp(
  mavispDB = NULL,
  proteins_of_interest = NULL,
  mode = "simple",
  simulation = "md"
)
}
\arguments{
\item{mavispDB}{path to the MAVISp database as a string}

\item{proteins_of_interest}{vector containing specific proteins of interest in HUGO format}

\item{mode}{string determining whether to use simple or ensemble mode of mavisp. Default is simple mode.
Takes values: 
\itemize{
\item simple
\item ensemble
}}

\item{simulations}{string specifying the simulation to use. The simulations possible can be found in the MAVISp database.}
}
\value{
returns a list of tibbles each containing the MAVISp entry of one protein. Each entry will contain an extra column specifying what data
the stability classification is based on.
}
\description{
This function loads the MAVISp database from the directory specified by the user.
}
\examples{

mavisp_data <- loadMAVISp(mavispDB = "/data/raw_data/computational_data/mavisp_database/biorxiv_v4_17112023",
          proteins_of_interest = c('NQO1','TP53'),
          mode = 'ensemble')

mavisp_data <- loadMAVISp(mavispDB = "/data/raw_data/computational_data/mavisp_database/biorxiv_v4_17112023") 
}
