% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadMAVISp.R
\name{loadMAVISp}
\alias{loadMAVISp}
\title{loadMAVISp}
\usage{
loadMAVISp(mavispDB = NULL, proteins_of_interest = NULL)
}
\arguments{
\item{mavispDB}{path to the MAVISp database}

\item{proteins_of_interest}{vector containing specific proteins of interest in HUGO format}
}
\value{
returns a list of tibbles each containing the MAVISp entry of one protein
}
\description{
This function loads the MAVISp database from the directory specified by the user.
}
\examples{

mavisp_data <- loadMAVISp(mavispDB = "/data/raw_data/computational_data/mavisp_database/biorxiv_v4_17112023",
          proteins_of_interest = c('NQO1','TP53'))

mavisp_data <- loadMAVISp(mavispDB = "/data/raw_data/computational_data/mavisp_database/biorxiv_v4_17112023") 
}
