% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TFinfluence.R
\name{TFinfluence}
\alias{TFinfluence}
\title{TFinfluence}
\usage{
TFinfluence(dataTRRUST, dataMAF, dataDEGs, dataMAVISp, dataTFexpr = FALSE)
}
\arguments{
\item{dataTRRUST}{RDA object of gene-TF pairs
Must contain the following columns:
\itemize{
\item TF (HUGO symbol of TF)
\item GENE (HUGO symbol of gene)
}}

\item{dataMAF}{A MAF file RDA object
The MAF file must at least contain the following columns:
\itemize{
\item Hugo_Symbol eg. BRCA1
\item HGVSp_Short eg. p.V83F
}}

\item{dataDEGs}{Output DEA function}

\item{dataMAVISp}{Output loadMAVISp function}
}
\value{
returns a tibble containing:
\itemize{
\item GENE
\item TF 
\item InteractionType 
\item PMID
\item tf_mutation
\item stab_class (the effect on stability as classified by MAVISp)
}
}
\description{
This function finds mutations in the transcription factors (TF) of the DEGs that have a TF in the database.
}
